Remove TODO from Private Key resource
Upon testing the "fastly_tls_private_key" resource with the "tls_private_key" one in the builtin TLS provider, it transpired that no string replacement was needed.
It seems terraform handles the string formatting for us.
However, the test still needs to do this as the SDK doesn't handle it, and the JSON formatting complains.