Don't cache dependencies in github PR workflow, instead rely on /vendor
Including the tfplugindocs module in vendor means it's updating with `go
mod vendor` along with the other libraries used. When running `go
install`, this vendored copy is used, and installed to a project-local
/bin directory. This enables the version of tfplugindocs used to be
independent of other go projects installed on one's system.

This change means `make dependencies` is no longer used, and isn't
needed in the github PR workflow. Furthermore, the source code for the
tool is included in the /vendor already so the caching of ~/go/* isn't
required either.