Fix dynamicsnippet content test
Testing TypeSet elements has changed in acceptance tests. Looking up a
preknown hash in the set no longer works, but there are new
TestCheckTypeSetElemAttr and TestCheckTypeSetElemNestedAttrs functions
for this instead which use "set.*" as the syntax.

Also the test was working incorrectly with a "count" attribute. I'm
guessing this worked before because a fake terraform implementation was
being used, but I couldn't get this working now. I tried ....content[0]
to try indexing the first instance of the "count"ed resource but it
still didn't work. In the end, I noticed that it was a bit redundant
anyway because there was always one thing being created, and the rest of
the fields would all have been duplicated if there were multiple.